okay these are not final changes by any means.
1. got rid of the extra resize image (seems to have imporved
performance)
2. started getting slider zooming to work (still in progress)
3. also in process of updated matrix entry structure and trying to make
it work with only non-zero vals

+++ src/Matrix2DEntry.java
@@ -3,9 +3,9 @@ public class Matrix2DEntry {
 	
 	private int genomeID;
 	private int pFamID;
-	private int count;
+	private short count;
 	
-	public Matrix2DEntry(int genomeID_in, int pFamID_in, int count_in) {
+	public Matrix2DEntry(int genomeID_in, int pFamID_in, short count_in) {
 		genomeID  = genomeID_in;
 		pFamID    = pFamID_in;
 		count     = count_in;
@@ -19,7 +19,7 @@ public class Matrix2DEntry {
 		return pFamID;
 	}
 	
-	public int getPFamCount() {
+	public short getPFamCount() {
 		return count;
 	}
 }
